requirements.txt
# Verified on conda env: py312
# LangChain Agent + FastAPI service
langchain==1.1.3
langchain-core==1.2.0
langchain-openai==1.0.1
langgraph==1.0.5
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.7.16

# HTTP 服务
fastapi==0.115.6
uvicorn[standard]==0.41.0

# 环境变量与配置
python-dotenv==1.2.1
pydantic-settings==2.6.1

# 测试
pytest==9.0.2
httpx==0.28.1

# 数据库
sqlalchemy==2.0.44
pymysql==1.1.2
cryptography==44.0.3
